requirements.txt
annotated-types==0.7.0
attrs==25.4.0
basemap==2.0.0
basemap_data==2.0.0
Cartopy==0.25.0
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
cyclopts==4.8.0
docstring_parser==0.17.0
docutils==0.22.4
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.2.6
numpy-stl==3.2.0
packaging==25.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pykdtree==1.4.3
pylance==2.0.1
pyparsing==3.3.2
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
pyvista==0.47.1
requests==2.32.5
rich==14.3.3
rich-rst==1.3.2
scooby==0.11.0
shapely==2.1.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vtk==9.6.0
